fix paths after moving assets to files folder
